Retired cube_manipulation.helper_functions using new utility.  All unit tests pass.
